Time: 1 ms (100%), Space: 44.5 MB (6.51%) - LeetHub

+++ 3733-minimum-time-to-complete-all-deliveries/3733-minimum-time-to-complete-all-deliveries.java
@@ -1,6 +1,6 @@
 class Solution {
     public long minimumTime(int[] d, int[] r) {
-        long left = 0, right = (long)(d[0] + d[1])*Math.max(r[0], r[1]);
+        long left = 0, right = ((long)d[0] + d[1])*Math.max(r[0], r[1]);
 
         while(left<right){
             long mid = left + (right-left)/2;
@@ -14,13 +14,13 @@ class Solution {
     }
 
     public boolean canFinish(int[] d, int[] r, long time){
-        long t1 = time - time/(long)r[0];
-        long t2 = time - time/(long)r[1];
+        long t1 = time - time/r[0];
+        long t2 = time - time/r[1];
 
-        long lcm = (long)r[0]*(long)r[1] / gcd((long)r[0], (long)r[1]);
+        long lcm = (long)r[0]*r[1] / gcd(r[0], r[1]);
         long t3 = time - time/lcm;
 
-        return t1>=(long)d[0] && t2>=(long)d[1] && t3>=(long)d[0] + (long)d[1];
+        return t1>=d[0] && t2>=d[1] && t3>=(long)d[0] + d[1];
     }
 
     public long gcd(long a, long b){
